prescription level adherence scores

+++ src/main/groovy/com/daugherty/pam/patient/Patient.groovy
@@ -28,5 +28,5 @@ class Patient {
   String ethnicity
   PatientPrescription[] prescriptions
   String notificationToken
-  Float adherence
+  Integer adherence
 }
